Rename http -> http_wrapper and update relative imports to absolute
- Renames http to http_wrapper to avoid conflicts with the third party library and allow for debugging
- Updates relative import to absolute to be more explicit and for easy file location updating
