+++ conjur/init/init_logic.py
@@ -16,7 +16,7 @@ import yaml
 
 from conjur.constants import DEFAULT_CONFIG_FILE
 from conjur.endpoints import ConjurEndpoint
-from conjur.http import invoke_endpoint, HttpVerb
+from conjur.http_wrapper import invoke_endpoint, HttpVerb
 
 DEFAULT_PORT = 443
