+++ conjur/login/login_logic.py
@@ -12,7 +12,7 @@ import logging
 
 # Internals
 from conjur.endpoints import ConjurEndpoint
-from conjur.http import invoke_endpoint, HttpVerb
+from conjur.http_wrapper import invoke_endpoint, HttpVerb
 
 class LoginLogic:
     """
